refactor(explore): move contentoutline types to separate file to remove circ deps
